Refactor the helpers for the repository to make the dependencies clearer
