+++ app/repositories/partnergraduateprogrammes/PartnerGraduateProgrammesRepository.scala
@@ -20,7 +20,7 @@ import model.Exceptions._
 import model.persisted.PartnerGraduateProgrammes
 import reactivemongo.api.DB
 import reactivemongo.bson.{ BSONDocument, _ }
-import repositories.BSONHelpers
+import repositories.ReactiveRepositoryHelpers
 import uk.gov.hmrc.mongo.ReactiveRepository
 import uk.gov.hmrc.mongo.json.ReactiveMongoFormats
 
@@ -36,7 +36,7 @@ trait PartnerGraduateProgrammesRepository {
 class PartnerGraduateProgrammesMongoRepository(implicit mongo: () => DB)
   extends ReactiveRepository[PartnerGraduateProgrammes, BSONObjectID]("application", mongo,
     PartnerGraduateProgrammes.partnerGraduateProgrammesFormat, ReactiveMongoFormats.objectIdFormats) with PartnerGraduateProgrammesRepository
-    with BSONHelpers {
+    with ReactiveRepositoryHelpers {
 
   val PartnerGraduateProgrammesCollection = "partner-graduate-programmes"
